pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yaml2pptx"
version = "0.1.0"
description = "Generate PowerPoint presentations from YAML files"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "python-pptx>=1.0.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
watch = [
    "watchdog>=4.0",
]
ai = [
    "anthropic>=0.40.0",
]

[project.scripts]
yaml2pptx = "yaml2pptx.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/yaml2pptx"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
